app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
alembic
psycopg2-binary
asyncpg
redis
celery[redis]
httpx
beautifulsoup4
google-generativeai
spacy
pypdf2
pandas
structlog
pytest
python-dotenv
python-multipart